Uploaded the code with associated files
Rplots contains all of the images that I used in the google doc. The salary.txt is the original file. Updated_salary.txt is just a file that I created to group by ID and create a salary_increment columns, which captures how much a person's salary has increased Y-o-Y